update textlogger usage
This change adds the configuration and command line flags setup for the
textlogger configuration. It is being added to address an error when
botting with the current code, that was changed to an updated klogr
during the kubernetes 1.29 update. The configuration code is inspired by
the machine-api-provider-vsphere main function.

Add the logtostderr flag to compensate for the logger changes.

Co-authored-by: Christopher J Schaefer <cjschaef@us.ibm.com>